download and deletion for groups. group deletion process is slightly different, still in progress

+++ src/server/api/account/index.ts
@@ -3,6 +3,7 @@ import Router from 'koa-router';
 import _ from 'lodash';
 import {Mutable} from 'utility-types';
 
+import {accountKeys} from '../_constants';
 import {knex} from '../../db/connection';
 import {encrypt, isValidPw, saltedHash, validateSchema} from '../../utils';
 import {deleteSchema, schema} from './_schema';
@@ -11,24 +12,37 @@ export const account = new Router();
 const route = '/api/v1/account';
 const table = 'accounts';
 
+
 account.delete(route, async (ctx: Koa.ParameterizedContext) => {
   const query: Mutable<tAccountQuery> = _.get(ctx, 'state.locals.data', {});
   const account: tAccount = _.get(ctx, 'state.user', {});
   await validateSchema<Mutable<tAccountQuery>>(ctx, deleteSchema, {
     ...query,
     id: account.id,
+    userId: account.userId,
   });
 
   const isValidPW = await isValidPw(query.password, account.password);
   if (!isValidPW) return ctx.throw(400, 'Password is not correct');
 
-  try {
-    await knex(table)
-      .limit(1)
-      .where({id: account.id})
-      .delete();
-  } catch (err) {
-    return ctx.throw(400, err);
+  // delete account and user references. leave rsvps, etc
+  if (account.userId) {
+    try {
+      await knex('users')
+        .limit(1)
+        .where({id: account.userId})
+        .delete();
+    } catch (err) {
+      return ctx.throw(400, err);
+    }
+    try {
+      await knex(table)
+        .limit(1)
+        .where({id: account.id})
+        .delete();
+    } catch (err) {
+      return ctx.throw(400, err);
+    }
   }
 
   ctx.body = {ok: true};
@@ -37,15 +51,9 @@ account.delete(route, async (ctx: Koa.ParameterizedContext) => {
 // TODO implement a POST route here as an upsert
 account.patch(route, async (ctx: Koa.ParameterizedContext) => {
   const query: Mutable<tAccountQuery> = _.get(ctx, 'state.locals.data', {});
+  await validateSchema<Mutable<tAccountQuery>>(ctx, schema, query);
+
   const loggedInAccount: tAccount = _.get(ctx, 'state.user', {});
-
-  try {
-    await schema.validateAsync(query);
-  } catch (err) {
-    const message = _.get(err, 'details[0].message', 'Bad Request');
-    return ctx.throw(400, message);
-  }
-
   const isValidPW = await isValidPw(query.password, loggedInAccount.password);
   if (!isValidPW) return ctx.throw(400, 'Password is not correct');
 
@@ -56,6 +64,10 @@ account.patch(route, async (ctx: Koa.ParameterizedContext) => {
 
   // password stuff will cause a constraint error - pull out before updating
   const {isFormSubmit, newPassword, password, ...updateQuery} = query;
+  console.log('updateQuery => ', updateQuery);
+  if (updateQuery.deletionDeadline === 'null') {
+    updateQuery.deletionDeadline = null;
+  }
 
   let updatedAccount: tAccount[];
   try {
@@ -63,7 +75,8 @@ account.patch(route, async (ctx: Koa.ParameterizedContext) => {
       .limit(1)
       .where({id: loggedInAccount.id})
       .update(updateQuery)
-      .returning('*');
+      // readonly string[] vs string[] trips typescript up
+      .returning(accountKeys as string[]);
   } catch (err) {
     return ctx.throw(400, err);
   }
